+++ Sources/Classes/Views/XibView/XibView.swift
@@ -26,9 +26,9 @@ import UIKit
 open class XibView: UIView {
 
     /// Nib name used to instantiate inner view
+    /// - NOTE: Be very carefully when you're intending to change this line
     open var innerViewNibName: String {
-        // Be very carefully when you're intending to change this line
-        return className(of: self)
+        return typeName(of: type(of: self))
     }
 
     public convenience init() {