+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - **Remove**: UICollectionView+CellRegistration
 - **Remove**: UIStoryboard+InstantiateViewController
 - **Remove**: `Array` extensions with `Set Algebra`
+- **Remove**: `NetworkService` extension for loading images
 
 ## 0.5.18
 - **Fix**: EmptyCell first appearance setup fix