+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - **Remove**: `Array` extensions with `Set Algebra`
 - **Remove**: `NetworkService` extension for loading images
 - **Remove**: `Observable` creation for `ImmutableMappable`
+- **Remove**: Drop `UIAlertController` extensions
 
 ## 0.5.18
 - **Fix**: EmptyCell first appearance setup fix